Reduced artifacts in the front ponies images (excluding the .jpg source file)
Through the use of one of Gimp's fantastic built-in features I was able
to remove most of the artifacts from the "front-ponies" series of
images. Look for yourself and see the changes!

Oh, and there are still a couple white speckles left over from the
original image, I'll go through the image manually and delete those
myself.